% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosmic_census.R
\name{cosmic_census}
\alias{cosmic_census}
\title{load COSMIC gene census}
\usage{
cosmic_census()
}
\description{
The COSMIC gene census list is just the genes
that the COSMIC project has curated. These are
all bona fide cancer genes. The csv file was
downloaded offline, since signin is required.
}
\examples{
cosmic_census()

}
\references{
See \url{https://cancer.sanger.ac.uk/census}
}
