% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_col_utils.R
\name{extract_polyphen}
\alias{extract_polyphen}
\title{Split a polyphen column into text and number}
\usage{
extract_polyphen(pphen)
}
\description{
This takes a column that has entries like
"Probably damaging(0.999)" and returns a data
frame with two columns containing the text
description and the number.
}
\examples{
testing = c("probably_damaging(0.999)")
extract_polyphen(testing)

}
