% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maf_col_utils.R
\name{extract_sift}
\alias{extract_sift}
\title{Split a SIFT column into text and number}
\usage{
extract_sift(sift)
}
\description{
This takes a column that has entries like
"deleterious(0.03)" and returns a data
frame with two columns containing the text
description and the number.
}
\examples{
testing = c("deleterious(0.03)")
extract_sift(testing)

}
