% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janeway_alterations.R
\name{get_janeway_alterations}
\alias{get_janeway_alterations}
\title{load Janeway alterations}
\usage{
get_janeway_alterations()
}
\value{
a tibble
}
\description{
The data from Proceedings of the National Academy of Sciences Dec 2014, 111 (51) E5564-E5573 are available in
what looks to be a relatively complete table that includes SNVs, SNPs, as well as copy number variation.
This function just loads those data for convenience. It is likely that further filtering is going to be
required before use.
}
\examples{
alts = get_janeway_alterations()
head(alts)
colnames(alts)

}
\references{
Data from \url{https://www.pnas.org/highwire/filestream/618044/field_highwire_adjunct_files/7/pnas.1419260111.sd08.xlsx}.
  See Proceedings of the National Academy of Sciences Dec 2014, 111 (51) E5564-E5573; DOI: 10.1073/pnas.1419260111.
}
