% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_janeway_snps.R
\name{get_janeway_snps}
\alias{get_janeway_snps}
\title{load Janeway PNAS SNVs}
\usage{
get_janeway_snps()
}
\value{
a tibble with variants and oncotator results
}
\description{
load Janeway PNAS SNVs
}
\examples{
snps = get_janeway_snps()
head(snps)
colnames(snps)
head(sort(table(snps$Hugo_Symbol), decreasing=TRUE)),25)

}
\references{
Data from \url{https://www.pnas.org/highwire/filestream/618044/field_highwire_adjunct_files/8/pnas.1419260111.sd09.xls}.
  See Proceedings of the National Academy of Sciences Dec 2014, 111 (51) E5564-E5573; DOI: 10.1073/pnas.1419260111.
}
