% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sleepingbeautydata.R
\docType{data}
\name{sleepingbeauty}
\alias{sleepingbeauty}
\title{Sleeping beauty gene list from Moriarty et al.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 372 rows and 7 columns.}
\source{
\url{https://dx.doi.org/10.1038/ng.3293}
}
\usage{
sleepingbeauty
}
\description{
A Sleeping Beauty forward genetic screen
identifies new genes and pathways driving
osteosarcoma development and metastasis.
}
\details{
This contains rows appended from Supplementary
tables 1-3.
}
\examples{
data(sleepingbeauty)

head(sleepingbeauty)


}
\keyword{data}
