% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_cnv_purity_corrected_calls.R
\name{target_cnv_calls}
\alias{target_cnv_calls}
\title{CNV calls, corrected for purity}
\usage{
target_cnv_calls()
}
\value{
A GRanges object describing regions of copy number change with
  medata columns:
  \itemize{
  \item{log2: log2 copy number}
  \item{baf: average B-Allele-Frequency for the copy number region}
  \item{cn: total integer copy number}
  \item{cn1: integer copy number of allele 1}
  \item{cn2: integer copy number of allele 2}
  \item{depth: average depth of region}
  \item{probes: from cnvKit, the number of bins}
  \item{weight: from cnvKit}
  }
}
\description{
This function loads the pureCN-corrected
copy number data as generated by cnvKit \code{call}.
The data were accessed from \code{/data/CCRBioinfo/projects/TargetOsteoDiscovery/report/summary.pureCN.cns.calls}
on December 13, 2018.
}
\examples{
x = target_cnv_calls()
head(x)
colnames(x)
# number of regions per sample
hist(as.vector(table(x$sampleName)))
hist(x$baf)
hist(x$log2)

}
