% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_qc_metrics.R
\name{target_dna_qc}
\alias{target_dna_qc}
\title{DNA sequencing QC metrics}
\usage{
target_dna_qc()
}
\description{
DNA sequencing QC metrics
}
\examples{
x = target_dna_qc()
head(x)
colnames(x)
hist(x$Percentage_duplication)
hist(x$PCT_TARGET_BASES_20X)
hist(x$PCT_NEAR_BAIT + x$PCT_ON_BAIT)

}
