% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/huex_data.R
\name{target_huex_se}
\alias{target_huex_se}
\title{Load TARGET huex data as SummarizedExperiment}
\usage{
target_huex_se()
}
\description{
Load TARGET huex data as SummarizedExperiment
}
\examples{

res = target_huex_se()

res

head(colData(res))
assay(res,'exprs')[1:5,1:5]
head(rowData(res))

}
