% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_data.R
\name{target_load_clinical}
\alias{target_load_clinical}
\title{load TARGET osteosarcoma data}
\usage{
target_load_clinical()
}
\description{
This loads the data from the local system.file
file. This file may or may not be the latest
version. The current verion is from 20180107.
}
\details{
The clinical columns include:

TARGET USI, Gender, Race, Ethnicity, Age at Diagnosis in Days, First Event, Time to First Event in Days, Vital Status, Overall Survival Time in Days, Year of Diagnosis, Year of Last Follow Up, Protocol, Disease at diagnosis, Metastasis site, Primary tumor site, Specific tumor site, Specific tumor side, Specific tumor region, Definitive Surgery, Primary site progression, Site of initial relapse, Time to first relapse in days, Time to first enrollment on relapse protocol in days, Time to first SMN in days, Time to death in days, Histologic response, Percent necrosis at Definitive Surgery, Relapse Type, Therapy, Comment, cohort
}
\examples{

target_load_clinical()

}
\references{
\url{ftp://caftpd.nci.nih.gov/pub/OCG-DCC/TARGET/OS/clinical/harmonized/}
}
