% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methylation_data.R
\name{target_methylation_se}
\alias{target_methylation_se}
\title{Load TARGET methylation data as SummarizedExperiment}
\usage{
target_methylation_se()
}
\description{
Load TARGET methylation data as SummarizedExperiment
}
