% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/purity_ploidy.R
\name{target_purity}
\alias{target_purity}
\title{Load Purity and Ploidy estimates}
\usage{
target_purity()
}
\value{
A \code{data.frame} with columns Source,
    Purity, Ploidy, and dataset
}
\description{
Results from PureCN run on discovery and
validation sets.
}
\examples{
pure = target_purity()
head(pure)
library(magrittr)
library(dplyr)
pure \%>\%
  group_by(dataset) \%>\%
  summarise(medploidy = median(Ploidy, na.rm=TRUE),
            medpurity = median(Purity, na.rm=TRUE),
            lowpurity = quantile(Purity, 0.1, na.rm=TRUE))
library(ggplot2)
p = ggplot(pure, aes(x=Purity))
  + geom_density(aes(group=dataset, color=dataset))
  + ggtitle('Tumor Purity Estimates')
print(p)

}
