% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_number_shiny.R
\name{target_shiny_copy_number_survival}
\alias{target_shiny_copy_number_survival}
\title{start copy-number/survival shiny app}
\usage{
target_shiny_copy_number_survival()
}
\description{
start copy-number/survival shiny app
}
