% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clinical_data.R
\name{target_usi_to_samplename}
\alias{target_usi_to_samplename}
\title{convert TARGET USI to samplename}
\usage{
target_usi_to_samplename(usi, long = FALSE)
}
\arguments{
\item{usi}{character() A target USI of the form "TARGET-40-ABCDEF" or "TARGET-40-ABCDEF-01A-10A"}

\item{long}{logical() Specify TRUE for USIs of the form "TARGET-40-ABCDEF-01A-10A" and
FALSE for "TARGET-40-ABCDEF"}
}
\description{
Convert from 'TARGET-40-PALPAX' to
'PALPAX'
}
